requirements.txt
kafka-python
confluent-kafka
mysql-connector-python